FIX Cantidad y Duración Subasta
Se agregó la cantidad y la duración en subasta con fecha inicio - fecha fin, se agregó el service de producto, y se acomodaron los exceptions

+++ AuctionMS.Application/Auction/Handlers/Commands/DeleteAuctionCommandHandler.cs
@@ -3,9 +3,8 @@ using AutoMapper;
 using AuctionMS.Core.Repository;
 using AuctionMS.Domain.Entities.Auction.ValueObjects;
 using AuctionMS.Application.Auction.Commands;
-using AuctionMS.Infrastructure.Repositories;
 using AuctionMS.Common.Dtos.Auction.Response;
-using AuctionMS.Infrastructure.Exceptions;
+using AuctionMS.Domain.Entities.Auction.ValueObjects;
 using AuctionMS.Core.RabbitMQ;
 
 namespace AuctionMS.Application.Auction.Handlers.Commands
@@ -25,21 +24,29 @@ namespace AuctionMS.Application.Auction.Handlers.Commands
 
         public async Task<Guid> Handle(DeleteAuctionCommand request, CancellationToken cancellationToken)
         {
-            try
+            if (request == null)
             {
-                var auctionId = AuctionId.Create(request.AuctionId);
-                var userId = AuctionUserId.Create(request.UserId);
-                var productId = AuctionProductId.Create(request.ProductId);
-                var auction = await _auctionRepository.GetByIdAsync(auctionId, userId, productId);
-                await _auctionRepository.DeleteAsync(auctionId);
-                var auctionDto = _mapper.Map<GetAuctionDto>(auction);
-                await _eventBus.PublishMessageAsync(auctionDto, "auctionQueue", "AUCTION_DELETED");
-                return auctionId.Value;
+                throw new ArgumentNullException(nameof(request), "Request cannot be null.");
             }
-            catch (Exception ex)
+
+
+            var auctionId = AuctionId.Create(request.AuctionId);
+            var userId = AuctionUserId.Create(request.UserId);
+            var productId = AuctionProductId.Create(request.ProductId);
+
+            var auction = await _auctionRepository.GetByIdAsync(auctionId, userId, productId);
+            if (auction == null)
             {
-                throw;
+                throw new Exception("Product not found."); // Esta excepción debe existir
             }
+
+            await _auctionRepository.DeleteAsync(auctionId);
+
+            var auctionDto = _mapper.Map<GetAuctionDto>(auction);
+            await _eventBus.PublishMessageAsync(auctionDto, "auctionQueue", "AUCTION_DELETED");
+
+            return auctionId.Value;
+
         }
     }
 }
